+++ src/command_modules/azure-cli-vm/HISTORY.rst
@@ -2,9 +2,10 @@
 
 Release History
 ===============
-
 2.0.18
 ++++++
+* vm : support `vm secret add/remove/list`
+* vm : `vm format-secret` is copied to `vm secret format`. The old one will be removed in future
 * Minor fixes.
 
 2.0.17